move determineTaskStatus to its own file
